Button class add

+++ src/entity/Entity.java
@@ -2,26 +2,27 @@
 package entity;
 
 public abstract class Entity {
-    protected double x, y;
-    public Entity(double x, double y){
+    protected int x, y;
+
+    public Entity(int x, int y) {
         this.x = x;
         this.y = y;
     }
 
-    public double getX() {
+    public int getX() {
         return x;
     }
 
-    public void setX(double x) {
+    public void setX(int x) {
         this.x = x;
     }
 
-    public double getY() {
+    public int getY() {
         return y;
     }
 
-    public void setY(double y) {
+    public void setY(int y) {
         this.y = y;
     }
-    
+
 }
